# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Recursos disponiveis para o aplicativo Smart POS"
  version: "1.0.0"
  title: "Linx Postos SmartPOS"
host: "postoshub.linx.com.br"
basePath: "/api"
schemes:
- "https"
paths:
...
  /v1/user/login/:
    post:
      tags:
      - "User"
      summary: "Login do usuário"
      description: "Realiza o login do usuário no backoffice"
      operationId: "services.user.web.post_user_login"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - in: "body"
        name: "data"
        description: "Dados do usuário"
        required: false
        schema:
          $ref: "#/definitions/data_1"
      responses:
        "200":
          description: "Login foi bem sucedido"
          schema:
            $ref: "#/definitions/inline_response_200_1"
      x-swagger-router-controller: "User"
  /v1/user/logout/:
    post:
      tags:
      - "User"
      summary: "Logout do usuário"
      description: "Realiza o logout do usuário no backoffice"
      operationId: "services.user.web.post_user_logout"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Logout foi bem sucedido"
      x-swagger-router-controller: "User"
  /v1/user/permission/:
    get:
      tags:
      - "User"
      summary: "Permissões do usuário"
      description: "Realiza a consulta de permissões do usuário no backoffice"
      operationId: "services.user.web.get_user_permission"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "username"
        in: "query"
        description: "Nome do usuário"
        required: false
        type: "string"
      - name: "password"
        in: "query"
        description: "Senha de acesso do usuário"
        required: false
        type: "string"
      - name: "rfid"
        in: "query"
        description: "Cartão RFID do usuário"
        required: false
        type: "string"
      responses:
        "200":
          description: "Consulta foi bem sucedida"
          schema:
            $ref: "#/definitions/inline_response_200_2"
      x-swagger-router-controller: "User"
...
definitions:
...
  data_1:
    type: "object"
    properties:
      username:
        type: "string"
        description: "Nome do usuário"
      password:
        type: "string"
        description: "Senha de acesso do usuário"
      rfid:
        type: "string"
        description: "Cartão RFID do usuário"
      force:
        type: "boolean"
        description: "Força o logout da sessão atual do usuário e cria uma nova"
  inline_response_200_1_permissions:
    type: "object"
    properties:
      saleCancel:
        type: "boolean"
        description: "Permissão para cancelar uma venda"
      withdrawal:
        type: "boolean"
        description: "Permissão para emitir uma sangria"
      supply:
        type: "boolean"
        description: "Permissão para emitir um suprimento"
      browseSales:
        type: "boolean"
        description: "Permissão para visualizar as vendas"
      changeSettings:
        type: "boolean"
        description: "Permissão para alterar a configuração do aplicativo"
    description: "Permissões de acesso do usuário"
    example:
      withdrawal: true
      browseSales: true
      supply: true
      saleCancel: true
      changeSettings: true
  inline_response_200_1:
    type: "object"
    properties:
      token:
        type: "string"
        description: "Token de acesso padrão JWT"
      salesmanName:
        type: "string"
        description: "Nome do vendedor associado ao usuário"
      permissions:
        $ref: "#/definitions/inline_response_200_1_permissions"
    example:
      salesmanName: "salesmanName"
      permissions:
        withdrawal: true
        browseSales: true
        supply: true
        saleCancel: true
        changeSettings: true
      token: "token"
  inline_response_200_2:
    type: "object"
    properties:
      salesmanId:
        type: "integer"
        description: "ID do vendedor associado ao usuário"
      salesmanName:
        type: "string"
        description: "Nome do vendedor associado ao usuário"
      permissions:
        $ref: "#/definitions/inline_response_200_1_permissions"
    example:
      salesmanName: "salesmanName"
      salesmanId: 0
      permissions:
        withdrawal: true
        browseSales: true
        supply: true
        saleCancel: true
        changeSettings: true
...